feat(resources): Полная интеграция всех типов ресурсов с backend хранилищем
Реализована полная интеграция 6 типов ресурсов (Effects, Filters, Transitions,
Templates, Style Templates, Subtitles) с backend state management.

## Backend изменения (Rust)

### Новые структуры ресурсов (project_state.rs)
- EffectResource - эффекты с параметрами
- FilterResource - фильтры
- TransitionResource - переходы
- TemplateResource - шаблоны раскладки
- StyleTemplateResource - стилистические шаблоны
- SubtitleResource - стили субтитров

### Расширение Project структуры
Добавлены 6 новых пулов в Project:
- effects_pool: HashMap<String, EffectResource>
- filters_pool: HashMap<String, FilterResource>
- transitions_pool: HashMap<String, TransitionResource>
- templates_pool: HashMap<String, TemplateResource>
- style_templates_pool: HashMap<String, StyleTemplateResource>
- subtitles_pool: HashMap<String, SubtitleResource>

### Новый модуль ResourceCommands (resources.rs)
Полная реализация команд для управления ресурсами:
- add_effect(), add_filter(), add_transition()
- add_template(), add_style_template(), add_subtitle()
- remove_resource() - универсальное удаление по типу

### Новые события (events.rs)
12 событий для отслеживания изменений ресурсов:
- EffectAdded/Removed, FilterAdded/Removed
- TransitionAdded/Removed, TemplateAdded/Removed
- StyleTemplateAdded/Removed, SubtitleAdded/Removed

### Интеграция в CommandHandler
- Добавлен ResourceCommands в handler.rs
- Маршрутизация SaveResource/DeleteResource по resource_type
- Публикация событий через EventBus

### Рефакторинг ImportedMedia
- Создан отдельный модуль imported_media.rs
- Исправлен borrow checker error в use_imported_media()
- Правильный порядок операций с state

## Frontend изменения (TypeScript)

### ResourcesProvider обновлен
Полная интеграция с backend state в resources-provider.tsx:
- Извлечение всех 6 типов из backend пулов
- Конвертация timestamp: секунды → миллисекунды
- Правильное преобразование HashMap → Array
- Сохранение структуры данных для каждого типа

### Обработка данных
- effectResources: извлечение из effects_pool
- filterResources: извлечение из filters_pool
- transitionResources: извлечение из transitions_pool
- templateResources: извлечение из templates_pool
- styleTemplateResources: извлечение из style_templates_pool
- subtitleResources: извлечение из subtitles_pool

## Документация

### Новый чек-лист
docs/ru/08_tasks/resources-panel-expansion-checklist.md:
- 8 фаз работы от backend до миграции
- Детальные задачи для каждой фазы
- Критерии завершения и метрики успеха

## Результаты

✅ Backend: 6 типов ресурсов хранятся в ProjectState
✅ Commands: Полная реализация add/remove для всех типов
✅ Events: 12 событий для отслеживания изменений
✅ Frontend: Извлечение всех ресурсов из backend state
✅ Type Safety: Правильная типизация на всех уровнях
✅ Build: Успешная сборка без ошибок

## Следующие шаги

- [ ] Тестирование полного цикла в dev режиме
- [ ] Генерация TypeScript bindings при запуске
- [ ] E2E тесты для операций с ресурсами
- [ ] Валидация данных на backend

+++ src-tauri/src/analysis/commands/content_commands.rs
@@ -229,8 +229,8 @@ pub async fn analyze_mood(
     primary: content_mood.mood.clone(),
     secondary: Vec::new(),
     valence: 0.0,
-    arousal: content_mood.energy_level as f64,
-    dominance: content_mood.emotional_intensity as f64,
+    arousal: content_mood.energy_level,
+    dominance: content_mood.emotional_intensity,
     emotional_arc: Vec::new(),
   };
 
@@ -309,8 +309,8 @@ pub async fn analyze_content_comprehensive(
     primary: content_mood.mood.clone(),
     secondary: Vec::new(),
     valence: 0.0,
-    arousal: content_mood.energy_level as f64,
-    dominance: content_mood.emotional_intensity as f64,
+    arousal: content_mood.energy_level,
+    dominance: content_mood.emotional_intensity,
     emotional_arc: Vec::new(),
   };
