+++ src-tauri/src/analysis/services/ai_director.rs
@@ -852,8 +852,8 @@ impl AIDirector {
               primary: content_mood.mood.clone(),
               secondary: Vec::new(),
               valence: 0.0,
-              arousal: content_mood.energy_level as f64,
-              dominance: content_mood.emotional_intensity as f64,
+              arousal: content_mood.energy_level,
+              dominance: content_mood.emotional_intensity,
               emotional_arc: Vec::new(),
             });
           }
@@ -865,8 +865,8 @@ impl AIDirector {
           primary: content_mood.mood.clone(),
           secondary: Vec::new(),
           valence: 0.0,
-          arousal: content_mood.energy_level as f64,
-          dominance: content_mood.emotional_intensity as f64,
+          arousal: content_mood.energy_level,
+          dominance: content_mood.emotional_intensity,
           emotional_arc: Vec::new(),
         });
       }
